Merges in the Permissions checks that I forgot to merge in last week.
Brian Gaffey

+++ Core/WPISuite/src/edu/wpi/cs/wpisuitetng/modules/core/entitymanagers/ProjectManager.java
@@ -28,11 +28,15 @@ import edu.wpi.cs.wpisuitetng.exceptions.ConflictException;
 import edu.wpi.cs.wpisuitetng.exceptions.DatabaseException;
 import edu.wpi.cs.wpisuitetng.exceptions.NotFoundException;
 import edu.wpi.cs.wpisuitetng.exceptions.NotImplementedException;
+import edu.wpi.cs.wpisuitetng.exceptions.UnauthorizedException;
 import edu.wpi.cs.wpisuitetng.exceptions.WPISuiteException;
+import edu.wpi.cs.wpisuitetng.Permission;
 import edu.wpi.cs.wpisuitetng.Session;
 import edu.wpi.cs.wpisuitetng.modules.EntityManager;
 import edu.wpi.cs.wpisuitetng.modules.Model;
 import edu.wpi.cs.wpisuitetng.modules.core.models.Project;
+import edu.wpi.cs.wpisuitetng.modules.core.models.Role;
+import edu.wpi.cs.wpisuitetng.modules.core.models.User;
 
 public class ProjectManager implements EntityManager<Project>{
 
@@ -58,7 +62,9 @@ public class ProjectManager implements EntityManager<Project>{
 	}
 
 	@Override
-	public Project makeEntity(Session s, String content) throws WPISuiteException {
+	public Project makeEntity(Session s, String content) throws WPISuiteException {	
+		User theUser = s.getUser();
+		if(theUser.getRole().equals(Role.ADMIN)){
 		
 		Project p;
 		try{
@@ -77,6 +83,10 @@ public class ProjectManager implements EntityManager<Project>{
 		}
 		
 		return p;
+		}
+		else{
+			throw new UnauthorizedException("You do not have enough priveldges to create a project.");
+		}
 	}
 
 	@Override
@@ -133,30 +143,47 @@ public class ProjectManager implements EntityManager<Project>{
 
 	@Override
 	public void save(Session s, Project model) throws WPISuiteException {
-		if(data.save(model))
-		{
-			return ;
+		User theUser = s.getUser();
+		if(model.getPermission(theUser).equals(Permission.WRITE) || 
+		   theUser.getRole().equals(Role.ADMIN)){
+			if(data.save(model))
+			{
+				return ;
+			}
+			else
+			{
+				throw new DatabaseException("Save failure for Project."); // Session User: " + s.getUsername() + " Project: " + model.getName());
+			}
 		}
 		else
-		{
-			throw new DatabaseException("Save failure for Project."); // Session User: " + s.getUsername() + " Project: " + model.getName());
-		}
+			throw new UnauthorizedException("You do not have the requred permissions to perform this action.");
 		
 	}
 
 	@Override
-	public boolean deleteEntity(Session s1, String id)
+	public boolean deleteEntity(Session s1, String id) throws WPISuiteException
 	{
-		
-		Model m = data.delete(data.retrieve(project, "idNum", id).get(0));
-		
-		return (m != null) ? true : false;
-		
+		User theUser = s1.getUser();
+		Project[] model = this.getEntity(id);
+		if(model[0].getPermission(theUser).equals(Permission.WRITE) || 
+		   theUser.getRole().equals(Role.ADMIN)){
+			Model m = data.delete(data.retrieve(project, "idNum", id).get(0));
+			
+			return (m != null) ? true : false;
+		}
+		else{
+			throw new UnauthorizedException("You do not have the required permissions to perform this action.");
+		}
 	}
 	
 	@Override
-	public void deleteAll(Session s) {
+	public void deleteAll(Session s) throws WPISuiteException {
+		User theUser = s.getUser();
+		if(theUser.getRole().equals(Role.ADMIN)){
 		data.deleteAll(new Project("",""));
+		}
+		else
+			throw new UnauthorizedException("You do not have the required permissions to perform this action.");
 	}
 
 	@Override
@@ -167,34 +194,39 @@ public class ProjectManager implements EntityManager<Project>{
 	
 	public Project update(Session s, Project toUpdate, String changeSet) throws WPISuiteException
 	{
-		// TODO: permissions checking here
+		User theUser = s.getUser();
+		if(toUpdate.getPermission(theUser).equals(Permission.WRITE) || 
+		   theUser.getRole().equals(Role.ADMIN)){
 		
-		// convert updateString into a Map, then load into the User
-		try
-		{
-			HashMap<String, Object> changeMap = new ObjectMapper().readValue(changeSet, HashMap.class);
-		
-			// check if the changeSet contains each field of User
-			if(changeMap.containsKey("name"))
+			// convert updateString into a Map, then load into the User
+			try
 			{
-				toUpdate.setName((String)changeMap.get("name"));
+				HashMap<String, Object> changeMap = new ObjectMapper().readValue(changeSet, HashMap.class);
+			
+				// check if the changeSet contains each field of User
+				if(changeMap.containsKey("name"))
+				{
+					toUpdate.setName((String)changeMap.get("name"));
+				}
+				
+				if(changeMap.containsKey("idNum"))
+				{
+					toUpdate.setIdNum((String)changeMap.get("idNum"));
+				}
+			}
+			catch(Exception e)
+			{
+				throw new DatabaseException("Failure in the ProjectManager.update() changeset mapper."); // on Mapping failure
 			}
 			
-			if(changeMap.containsKey("idNum"))
-			{
-				toUpdate.setIdNum((String)changeMap.get("idNum"));
-			}
+			// save the changes back
+			this.save(s, toUpdate);
+			
+			// check for changes in each field
+			return toUpdate;
 		}
-		catch(Exception e)
-		{
-			throw new DatabaseException("Failure in the ProjectManager.update() changeset mapper."); // on Mapping failure
-		}
-		
-		// save the changes back
-		this.save(s, toUpdate);
-		
-		// check for changes in each field
-		return toUpdate;
+		else
+			throw new UnauthorizedException("You do not have the required permissions to perform this action.");
 	}
 
 	@Override
